# Allow controlling time flow for EmbeddedEventLoop