--- a/transport/src/main/java/io/netty/channel/embedded/EmbeddedEventLoop.java
+++ b/transport/src/main/java/io/netty/channel/embedded/EmbeddedEventLoop.java
@@ -30,6 +30,22 @@ import java.util.Queue;
 import java.util.concurrent.TimeUnit;
 
 final class EmbeddedEventLoop extends AbstractScheduledEventExecutor implements EventLoop {
+    /**
+     * When time is not {@link #timeFrozen frozen}, the base time to subtract from {@link System#nanoTime()}. When time
+     * is frozen, this variable is unused.
+     *
+     * Initialized to {@link #initialNanoTime()} so that until one of the time mutator methods is called,
+     * {@link #getCurrentTimeNanos()} matches the default behavior.
+     */
+    private long startTime = initialNanoTime();
+    /**
+     * When time is frozen, the timestamp returned by {@link #getCurrentTimeNanos()}. When unfrozen, this is unused.
+     */
+    private long frozenTimestamp = 0;
+    /**
+     * Whether time is currently frozen.
+     */
+    private boolean timeFrozen = false;
 
     private final Queue<Runnable> tasks = new ArrayDeque<Runnable>(2);
 
@@ -60,7 +76,7 @@ final class EmbeddedEventLoop extends AbstractScheduledEventExecutor implements 
     }
 
     long runScheduledTasks() {
-        long time = AbstractScheduledEventExecutor.nanoTime();
+        long time = getCurrentTimeNanos();
         for (;;) {
             Runnable task = pollScheduledTask(time);
             if (task == null) {
@@ -73,6 +89,41 @@ final class EmbeddedEventLoop extends AbstractScheduledEventExecutor implements 
 
     long nextScheduledTask() {
         return nextScheduledTaskNano();
+    }
+
+    @Override
+    protected long getCurrentTimeNanos() {
+        if (timeFrozen) {
+            return frozenTimestamp;
+        } else {
+            return System.nanoTime() - startTime;
+        }
+    }
+
+    void advanceTimeBy(long nanos) {
+        if (timeFrozen) {
+            frozenTimestamp += nanos;
+        } else {
+            // startTime is subtracted from nanoTime, so increasing the startTime will advance getCurrentTimeNanos
+            startTime -= nanos;
+        }
+    }
+
+    void freezeTime() {
+        if (!timeFrozen) {
+            frozenTimestamp = getCurrentTimeNanos();
+            timeFrozen = true;
+        }
+    }
+
+    void unfreezeTime() {
+        if (timeFrozen) {
+            // we want getCurrentTimeNanos to continue right where frozenTimestamp left off:
+            // getCurrentTimeNanos = nanoTime - startTime = frozenTimestamp
+            // then solve for startTime
+            startTime = System.nanoTime() - frozenTimestamp;
+            timeFrozen = false;
+        }
     }
 
     @Override