--- a/transport/src/main/java/io/netty/channel/embedded/EmbeddedEventLoop.java
+++ b/transport/src/main/java/io/netty/channel/embedded/EmbeddedEventLoop.java
@@ -95,6 +95,7 @@ final class EmbeddedEventLoop extends AbstractScheduledEventExecutor implements 
     protected long getCurrentTimeNanos() {
         if (timeFrozen) {
             return frozenTimestamp;
+        }
         return System.nanoTime() - startTime;
     }