@chrisvest 2022-06-10
> Note that if an EmbeddedEventLoop is present in the application, the performance impact is likely substantially higher, because this would necessitate a virtual call. However this is not an issue for production applications, and the affected code is still not very hot.

The `EmbeddedChannel` is used in the compression codecs. However, the morphism is a property of the individual call-sites, so if every call-site ends up inlining and monomorphise this call, then it still won't add any virtual call overhead.