@chrisvest 2022-06-10
Yes, that matches my understanding as well.