@yawkat 2022-06-13
ive run the benchmark again with a `new EmbeddedChannel();` before the trials on the same jvm, and the difference is about 1%. I assume any trap required for the monomorphic call inlining is merged with the cast check from `scheduledExecutorService()` or something. But since the difference is so tiny I won't investigate the assembly.